requirements.txt
numpy==1.21.5
deepface==0.0.75
insightface==0.6.2
mediapipe==0.8.10.1
mtcnn==0.1.1
#numpy==1.21.5
Pillow==9.2.0
retinaface==1.1.1
scikit_image==0.19.2

